Still refactorying my code, made so many mistakes before
I am still testing things around, trying to clean it as much as i can.
Since I am tyring to implement delta time for a smooth and balance speed
of the ball movement independent from the the framerate.

It fixed the speed and its consisten, which means it does not change if
the framerate changed. But, after changing the Point to float and
casting it,  its movement does not look good but it works.