some refactoring and adding the animation
animation of the paddle when moving left and right.

So we have jumpby, speed, targetX and a moving bool. To help ensure the
movement smooth that depeneds on delta or the the it takes on each frame
we get.

Basically, once the paddle reach the target, the movement of should
stop.